modified:   index.html modified:   main.js
Renamed "reset" to "clear" for clarity in the tool section.
Reset will be a name for the new function that resets the settings

Expanding to have 2 state of management. One for tools, and the other for the color

Next step: create the function for managing the state of the color