modified:   index.html modified:   main.js modified:   styles.css
Created a slider to get the value for the canvas.

Separated the square creating function to createSquares().
Renamed addDiv() to addToCanvas(). Adjusted so it can dynamically get the slider value.

Next Step: currently, the function clear() and toggleBorder() do not work due to it being a static event listener.
Change it so it uses event delegation.