modified:   main.js modified:   styles.css
Event delegation was not the problem with clear() and toggleBorder().
It had something to do with the timing of adding the event listener.

Updated clear() by moving the declaration  of the square variable to inside of the button event listener.

Updated toggleBorder() to target the parent element (which is the canvas),
and not targeting each of the squares.